requirements.txt
django-5.0.6
Faker==25.8.0
psycopg2==2.9.9
